correcciones

+++ src/main/java/mx/edut/uteq/evaluacion1/model/repository/CategoriaRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import mx.edut.uteq.evaluacion1.model.entity.Categoria;
 
-public interface CategoriaRepo extends JpaRepository<Categoria,Long>{
+public interface CategoriaRepo extends JpaRepository<Categoria,Integer>{
     
 }
